The current stm32 drivers will only enable soft NSS support if the device is using a Zephyr managed GPIO for CS, or if the CONFIG_SPI_STM32_USE_HW_SS flag is set. Neither of these are actually appropriate for a peripheral, which doesn't control CS.
With this change, it is possible for the SPI interface to be
configured as a peripheral but still use soft NSS -- otherwise,
short of creating a "fake" GPIO configuration, the peripheral
will always use hard NSS, and for some boards/configurations,
that is unworkable. In the case of a board where the CS is not
hooked up, we can use soft NSS to allow the peripheral to assume
CS and be driven by the SPI clock signal alone.

Our board has only three signals (clock, COPI, POCI) enabled on
one of the SPI interfaces (it is communicating with another MCU
on the same board), and works with this change. Without the
change, we must enable a "fake" GPIO driver configuration (which
is never used for anything, and really makes little sense for a
peripheral, but it just happens to force the driver to go down
that code path). We cannot disable the CONFIG_SPI_STM32_USE_HW_SS
mechanism, as that impacts ALL interfaces, and we have a second
interface in controller mode that speaks to flash and requires
the NSS_HARD codepath.